Updated style.css and index.html
Setup the page with media queries to change format at 768px. Got all of it ready to go, except the footer. The list items reverse order when I float them to the right. The CSS is a little hacked up at the moment trying things out. I'm walking away to give myself time to think about it.